"Add middleware for URL-encoded form data"
This commit adds a new middleware, `express.urlencoded({ extended: true })`, to the Express application. This middleware will parse incoming requests with URL-encoded payloads (such as those sent by HTML forms) and make the data available on `req.body`. The `extended` option is set to `true` to enable parsing of nested objects in the URL-encoded data.