"Refactored error handling middleware and added custom error class; used custom error in task controller methods"
This commit includes changes to both the `controllers/tasks.js` file and the `middleware/error-handler.js` file. In the `tasks.js` file, several error responses have been updated to use a new `createCustomError` function, which has been imported from `../errors/custom-error`. In the `error-handler.js` file, a new `CustomAPIError` class has been defined, and the `errorHandlerMiddleware` function has been updated to check if an error is an instance of `CustomAPIError` and handle it accordingly.