"Add product querying functionality to controllers/products.js; export two controller functions for getting all products"
This commit adds functionality to the products controller file in a serverless API. It includes code for querying products based on various parameters such as featured, company, name, sorting, and fields. Two controller functions are added: getAllProductsStatic and getAllProducts, both of which are used to fetch all products with different querying options.